Portões de autorização

+++ app/Http/Controllers/UsersController.php
@@ -12,64 +12,82 @@ use Inertia\Inertia;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Request;
 use Illuminate\Support\Facades\Redirect;
+use Illuminate\Support\Facades\Gate;
 
 class UsersController extends Controller
 {
     public function index()
     {
-        return Inertia::render('Users/Index', [
-            'filters' => Request::all('search', 'role', 'trashed'),
-            'users' => new UserCollection(
-                Auth::user()->account->users()
-                    ->orderByName()
-                    ->filter(Request::only('search', 'role', 'trashed'))
-                    ->paginate()
-                    ->appends(Request::all())
-            ),
-        ]);
+        $response = Gate::inspect('isAdmin');
+        if ($response->allowed()) {
+            return Inertia::render('Users/Index', [
+                'filters' => Request::all('search', 'role', 'trashed'),
+                'users' => new UserCollection(
+                    Auth::user()->account->users()
+                        ->orderByName()
+                        ->filter(Request::only('search', 'role', 'trashed'))
+                        ->paginate()
+                        ->appends(Request::all())
+                ),
+            ]);
+        }
     }
 
     public function create()
     {
-        return Inertia::render('Users/Create');
+        $response = Gate::inspect('isAdmin');
+        if ($response->allowed()) {
+            return Inertia::render('Users/Create');
+        }
     }
 
     public function store(UserStoreRequest $request)
     {
-        Auth::user()->account->users()->create(
-            $request->validated()
-        );
-
-        return Redirect::route('users')->with('success', 'Utilizador criado.');
+        $response = Gate::inspect('isAdmin');
+        if ($response->allowed()) {
+            Auth::user()->account->users()->create(
+                $request->validated()
+            );
+            return Redirect::route('users')->with('success', 'Utilizador criado.');
+        }
     }
 
     public function edit(User $user)
     {
-        return Inertia::render('Users/Edit', [
-            'user' => new UserResource($user),
-        ]);
+        $response = Gate::inspect('isAdmin');
+        if ($response->allowed()) {
+            return Inertia::render('Users/Edit', [
+                'user' => new UserResource($user),
+            ]);
+        }
     }
 
     public function update(User $user, UserUpdateRequest $request)
     {
-        $user->update(
-            $request->validated()
-        );
-
-        return Redirect::back()->with('success', 'Utilizador actualizado.');
+        $response = Gate::inspect('isAdmin');
+        if ($response->allowed()) {
+            $user->update(
+                $request->validated()
+            );
+            return Redirect::back()->with('success', 'Utilizador actualizado.');
+        }
     }
 
     public function destroy(User $user, UserDeleteRequest $request)
     {
-        $user->delete();
-
-        return Redirect::back()->with('success', 'Utilizador eliminado.');
+        $response = Gate::inspect('isAdmin');
+        if ($response->allowed()) {
+            $user->delete();
+            return Redirect::back()->with('success', 'Utilizador eliminado.');
+        }
     }
 
     public function restore(User $user)
     {
-        $user->restore();
-
-        return Redirect::back()->with('success', 'Utilizador restaurado.');
+        $response = Gate::inspect('isAdmin');
+        if ($response->allowed()) {
+            $user->restore();
+            return Redirect::back()->with('success', 'Utilizador restaurado.');
+        }
     }
 }
